refactored cisd derivatives
