+++ pycc/cideriv.py
@@ -1,20 +1,37 @@
 """
 cideriv.py: CISD analytic-derivative property driver.
 
-`CIderiv` is the CISD leaf of the :class:`~pycc.correlatedderivs.CorrelatedDerivs` hierarchy.
+`CIderiv` is the CISD leaf of the :class:`~pycc.correlatedderivs.CorrelatedDerivs` hierarchy,
+matching the `MPwfn`/`MPderiv` and `CCwfn`/`CCderiv` split: `CIwfn` holds only wavefunction
+quantities (amplitude solve, energy, raw densities); this class supplies the two density hooks
+(`_unrelaxed_densities`, `_perturbed_unrelaxed_densities`) and inherits `relaxed_dipole`,
+`gradient`, `polarizability`, `dipole_derivatives`, and `hessian` from the base unchanged. Only
+the genuinely CISD-specific wave-function-overlap properties - `atomic_axial_tensors` and
+`velocity_dipole_derivatives` - are custom code, together with the coupled-perturbed-CI (CPCI)
+response machinery they (and the two density hooks) are built from.
 
-* `relaxed_dipole` and `gradient` are inherited from the base, driven by the two density hooks
-  below (:meth:`_unrelaxed_densities` / :meth:`_perturbed_unrelaxed_densities`, built from
-  `CIwfn._corr_QGX()`'s true-normalized correlation 1-/2-PDM).
-* `polarizability`, `dipole_derivatives`, and `hessian` are custom, using `CIwfn`'s own
-  equilibrium Z-vector (`_zvector`/`_corr_QGX`/`_solve_dz_dR`) rather than the base's generic
-  machinery.
-* `atomic_axial_tensors` and `velocity_dipole_derivatives` are custom wave-function-overlap
-  constructions the base does not provide at all.
+DENSITY HOOKS
+-------------
+Both hooks return the correlation 1-PDM and the Klein-four-symmetrized PURE cumulant 2-PDM (a
+0.25*(4-permutation) average, NO correlation-density cross term, no HF block -- see
+`_cisd_symmetrize` for the convention derivation), true-normalized. The base's generalized-Fock
+Lagrangian is built on the Fock-convention partitioning E_corr = Tr(D f) + Tr(Gam ERI) (like
+MP2's densities); the mean-field coupling of the 1-PDM is supplied by the Lagrangian's own
+termA/termB, not folded into Gam.
 
-CPCI/magnetic/vecpot machinery (`_cpci_ints`/`_solve_cpci`/`_cpci_raw`, the magnetic/vecpot
-integral builders) and the density builders (`_perturbed_cisd_corr_opdm`/`_perturbed_cisd_tpdm`/
-`_corr_QGX`/`_zvector`) stay on `CIwfn` and are called from here.
+`_perturbed_unrelaxed_densities` solves the CPCI response directly from the base's own canonical,
+full-occupied-space perturbed integrals (`df`, `deri`) via `_solve_cpci_ints` -- the CISD analog of
+`CCderiv._perturbed_amplitudes`. This is the piece that used to be dead code: the base's second-
+derivative path was previously bypassed by CISD-specific overrides of `dipole_derivatives`/`hessian`/
+`polarizability`, which have been removed in favor of the base's inherited implementations.
+
+CPCI/MAGNETIC/VECPOT MACHINERY
+-------------------------------
+`_cpci_ints`/`_solve_cpci`/`_solve_cpci_ints`/`_cpci_raw`, the magnetic/vector-potential integral
+builders, and the raw perturbed-density builders live here (not on `CIwfn`) - they are about
+derivatives, not the wavefunction. `_solve_cpci_ints` is the core iterative solve, taking already-
+built perturbed integrals directly; `_solve_cpci` is the `Perturbation`-keyed, cached entry point
+used by the AAT/VG-APT overlap code (magnetic/vecpot/nuclear, via `_cpci_ints`).
 
 """
 
@@ -26,476 +43,425 @@ from .correlatedderivs import CorrelatedDerivs
 
 
 class CIderiv(CorrelatedDerivs):
-    """CISD correlation derivative-property driver. Constructed from a converged CIwfn
-    (aliased `self.ci`). See the module docstring for which properties are inherited vs custom.
-    Spatial (closed-shell RHF), all-electron only."""
+    """CISD correlation derivative-property driver. Constructed from a converged CIwfn (aliased
+    `self.ci`). `relaxed_dipole`/`gradient`/`polarizability`/`dipole_derivatives`/
+    `hessian` are inherited from `CorrelatedDerivs`, driven by the two density hooks below.
+    `atomic_axial_tensors`/`velocity_dipole_derivatives` are custom wave-function-overlap
+    constructions. Spatial (closed-shell RHF), all-electron only."""
 
     def __init__(self, ciwfn) -> None:
         super().__init__(ciwfn)
         self.ci = ciwfn                                # alias: this class uses .ci, the base .wfn
 
+    def _cisd_symmetrize(self, D_pqrs_raw):
+        return 0.25 * (D_pqrs_raw + D_pqrs_raw.transpose(1, 0, 3, 2)
+                       + D_pqrs_raw.transpose(2, 3, 0, 1) + D_pqrs_raw.transpose(3, 2, 1, 0))
+
     def _unrelaxed_densities(self):
-        """(D, Gam) from CIwfn._corr_QGX(). Feeds relaxed_dipole/gradient only."""
-        Q_corr, G_corr, _X_corr = self.ci._corr_QGX()
-        return Q_corr, G_corr
+        """CISD unrelaxed reduced densities (D, Gam) as full-MO arrays."""
+        _D_pq, D_pq_corr, D_pqrs = self.ci._cisd_densities()
+        Gam = self._cisd_symmetrize(D_pqrs)
+        return D_pq_corr, Gam
 
-    def _perturbed_unrelaxed_densities(self, pert, df=None, deri=None, dL=None):
-        """First-order response (d_x D, d_x Gam) to pert. Feeds base-driven gradient/relaxed_dipole
-        second-derivative uses; dipole_derivatives/hessian/polarizability use their own _solve_dz_dR below instead."""
-        ci = self.ci
-        no = ci.no
-        dc1, dc2, dc0v = ci._solve_cpci(pert)
-
-        dD_corr = ci._perturbed_cisd_corr_opdm(dc1, dc2)
-        dG_raw = ci._perturbed_cisd_tpdm(dc1, dc2, dc0v)
-
-        dG_sym = 0.25 * (dG_raw + dG_raw.transpose(1, 0, 3, 2) + dG_raw.transpose(2, 3, 0, 1) + dG_raw.transpose(3, 2, 1, 0))
-        dGam = dG_sym.copy()
-        dG_cross = np.zeros_like(dGam)
-        for i in range(no):
-            dG_cross[i, :, i, :] += 2.0 * dD_corr
-            dG_cross[:, i, :, i] += 2.0 * dD_corr
-            dG_cross[i, :, :, i] -= dD_corr.T
-            dG_cross[:, i, i, :] -= dD_corr
-        dGam = dGam + 0.5 * dG_cross
-
+    def _perturbed_unrelaxed_densities(self, pert, df, deri, dL):
+        """First-order response (d_x D, d_x Gam) of the CISD unrelaxed reduced densities to
+        pert, in the same convention as _unrelaxed_densities. Solves the coupled-perturbed-CI
+        response directly from the base's own canonical, full-occupied-space perturbed integrals. """
+        dc1, dc2, dc0v, _dt1, _dt2 = self._solve_cpci_ints(df, deri, imaginary=False)
+        dD_corr = self._perturbed_cisd_corr_opdm(dc1, dc2)
+        dG_raw = self._perturbed_cisd_tpdm(dc1, dc2, dc0v)
+        dGam = self._cisd_symmetrize(dG_raw)
         return dD_corr, dGam
 
-    # length-gauge APT 
+    # coupled-perturbed-CI response 
 
-    def polarizability(self, route: str = '2n+1') -> np.ndarray:
-        """CISD correlation static polarizability (a.u.), shape (3, 3):
-        alpha[a,b] = -d^2 E_corr / dF_a dF_b. Both routes are custom,
-        using CIwfn's own equilibrium Z-vector.
+    def _psi4_mints(self):
+        """psi4 MintsHelper + C as psi4.core.Matrix"""
+        if getattr(self, '_mints_cache', None) is None:
+            import psi4
+            mints = psi4.core.MintsHelper(self.ci.H.basisset)
+            C_p4 = psi4.core.Matrix.from_array(np.asarray(self.ci.C))
+            self._mints_cache = (mints, C_p4)
+        return self._mints_cache
 
-        route='2n+1' (default): differentiate the relaxed dipole a second
-        time. See _polarizability_2n1().
-        route='explicit': differentiate the energy twice, reusing the
-        Hessian's T0-T5/Z-vector machinery for field-field pairs. See
-        _polarizability_explicit()."""
-        if route == '2n+1':
-            return self._polarizability_2n1()
-        if route == 'explicit':
-            return self._polarizability_explicit()
-        raise ValueError(f"unknown polarizability route {route!r} (use '2n+1' or 'explicit')")
-
-    def _polarizability_2n1(self) -> np.ndarray:
-        """2n+1 route: uses CIwfn's own equilibrium Z-vector and perturbed
-        Z-vector solve, matching dipole_derivatives()/hessian()."""
-        from .cphf import Perturbation
-        c = self.contract
-        ci = self.ci
-        o, v = ci.o, ci.v
-
-        zv = ci._zvector()
-        Drel = zv['Q_relaxed'].copy()
-        for i in range(ci.no):
-            Drel[i, i] -= 2.0
-        mu = [np.asarray(ci.H.mu[a]) for a in range(3)]
-
-        alpha = np.zeros((3, 3))
-        for b in range(3):
-            pert = Perturbation('field', b)
-            U_b = np.asarray(ci.cphf.full_U(pert))
-            dc1, dc2, dc0v = ci._solve_cpci(pert)
-            dc1, dc2 = dc1.real, dc2.real
-            dc0v = dc0v.real if hasattr(dc0v, 'real') else dc0v
-            dz, dD_corr = self._solve_dz_dR(pert, dc1, dc2, dc0v)
-            dDrel = dD_corr.copy()
-            dDrel[v, o] += dz
-            dDrel[o, v] += dz.T
-            for a in range(3):
-                rot = U_b.T @ mu[a] + mu[a] @ U_b
-                alpha[a, b] = (c('pq,pq->', dDrel, mu[a]) + c('pq,pq->', Drel, rot))
-        return alpha
-
-    def _polarizability_explicit(self) -> np.ndarray:
-        """Explicit route: reuses the Hessian's Z-vector machinery
-        (build_xi_ab/build_X_deriv/build_cpci_term/_hess_build_Y/_hess_X_tilde)
-        for field-field pairs - all second-derivative skeletons vanish
-        for a static field, so no U^{fg} solve is needed."""
-        from .cphf import Perturbation
+    def _build_magnetic_ints(self, beta):
         ct = self.contract
-        ci = self.ci
-        o, v, no, nv = ci.o, ci.v, ci.no, ci.nv
-        F = np.asarray(ci.H.F)
-        eps = np.diag(F)
-        nmo = ci.nmo
-
-        zv = ci._zvector()
-        Q, G, _ = ci._corr_QGX()
-        X_t = self._hess_X_tilde()
-        Y = self._hess_build_Y()
-        ERI = np.asarray(ci.H.ERI)
-        z = zv['z_ai']
-        n0 = ci._normalized_amplitudes()[0]
-
-        ERI_M = ERI.transpose(0, 2, 1, 3)
-        W_M = 2 * ERI_M - ERI_M.transpose(0, 2, 1, 3)
-        A_M = 4 * ERI_M - ERI_M.transpose(0, 2, 1, 3) - ERI_M.transpose(0, 3, 2, 1)
-
-        mu = [np.asarray(ci.H.mu[a]) for a in range(3)]
-        zero_g = np.zeros((nmo, nmo, nmo, nmo))
-        zero_S = np.zeros((nmo, nmo))
-
-        # First-order field quantities: U^a, field-CPCI raw amplitudes,
-        # X_deriv[a] = -Q_corr.mu_a, Fa_M[a] = -mu_a (g-dependent piece is 0)
-        U_all, dT1_all, dT2_all, X_deriv, Fa_M = {}, {}, {}, {}, {}
-        for a in range(3):
-            pert = Perturbation('field', a)
-            U_all[a] = np.asarray(ci.cphf.full_U(pert))
-            dT1_all[a], dT2_all[a] = ci._cpci_raw(pert)
-            h_a_skel = -mu[a]
-            X_deriv[a] = self.build_X_deriv(h_a_skel, zero_g)
-            Fa_M[a] = h_a_skel.copy()
-            # Aa_M[a] is identically zero for a field (built from g^f = 0)
-
-        alpha = np.zeros((3, 3))
-        for a in range(3):
-            for b in range(3):
-                U_a, U_b = U_all[a], U_all[b]
-
-                # T0: skeleton density contraction - vanishes (h^{fg}=g^{fg}=0)
-                t0_val = 0.0
-
-                # T2: Z-vector replacement (no U^{fg} solved - same as hessian())
-                xi = self.build_xi_ab(U_a, U_b, zero_S, zero_S, zero_S)
-                t2_xiXt = -ct('ij,ji->', xi, X_t)
-                # F_ab_M = 0 (h^{fg} = g^{fg} = 0)
-                Bab = np.zeros((nmo, nmo))
-                Bab -= ct('ij,j->ij', xi, eps)
-                Bab -= ct('kl,ijkl->ij', xi[o, o], W_M[:, :, o, o])
-                Bab += ct('ki,kj->ij', U_a, Fa_M[b])
-                Bab += ct('ki,kj->ij', U_b, Fa_M[a])
-                Bab += ct('kj,ik->ij', U_a, Fa_M[b])
-                Bab += ct('kj,ik->ij', U_b, Fa_M[a])
-                Bab += ct('ki,kj,k->ij', U_a, U_b, eps)
-                Bab += ct('ki,kj,k->ij', U_b, U_a, eps)
-                UaUbT = U_a[:, o] @ U_b[:, o].T
-                Bab += ct('kl,ijkl->ij', UaUbT, A_M)
-                temp5_b = ct('lm,kjlm->kj', U_b[:, o], A_M[:, :, :, o])
-                temp5_a = ct('lm,kjlm->kj', U_a[:, o], A_M[:, :, :, o])
-                Bab += ct('ki,kj->ij', U_a, temp5_b)
-                Bab += ct('ki,kj->ij', U_b, temp5_a)
-                temp6_b = ct('lm,iklm->ik', U_b[:, o], A_M[:, :, :, o])
-                temp6_a = ct('lm,iklm->ik', U_a[:, o], A_M[:, :, :, o])
-                Bab += ct('kj,ik->ij', U_a, temp6_b)
-                Bab += ct('kj,ik->ij', U_b, temp6_a)
-                # Aa_M[a]/[b] terms dropped - identically zero for a field
-                t2_zvec = 2.0 * ct('ai,ai->', Bab[v, o], z)
-                t2_val = t2_xiXt + t2_zvec
-
-                # T3+T4: orbital-response Lagrangian + Y tensor (Y equilibrium, reused)
-                t3_val = 2.0 * (ct('ij,ij->', U_b, X_deriv[a])
-                                + ct('ij,ij->', U_a, X_deriv[b]))
-                t4_val = 2.0 * ct('ij,kl,ijkl->', U_a, U_b, Y)
-
-                # T5: CPCI amplitude response (field-perturbed raw amplitudes)
-                t5_val = n0**2 * self.build_cpci_term(
-                    dT1_all[a], dT2_all[a], dT1_all[b], dT2_all[b]).real
-
-                # alpha = -d^2 E_corr / dF dG (hessian() returns +d^2E/dXdY
-                # with no extra sign, so the field-field analog needs the
-                # explicit minus here)
-                alpha[a, b] = -(t0_val + t2_val + t3_val + t4_val + t5_val).real
-
-        alpha = 0.5 * (alpha + alpha.T)
-        return alpha
-
-    def dipole_derivatives(self, route: str = 'explicit') -> np.ndarray:
-        """CISD CORRELATION-ONLY LG-APT, shape (natom, 3, 3), [A, beta, alpha].
-        """
-        from .cphf import Perturbation
-        c = self.contract
+        nbf = self.ci.nmo
         o, v = self.ci.o, self.ci.v
-        zv = self.ci._zvector()
-        # Correlation-only relaxed density: Q_relaxed minus the HF 2*delta_oo.
-        # T3/T3z are already pure correlation (amplitude/Z-vector response).
-        Qt = zv['Q_relaxed'].copy()
-        for i in range(self.ci.no):
-            Qt[i, i] -= 2.0
-        natom = self.ci.derivatives.natom
-        h_E = [np.asarray(self.ci.H.mu[f]) for f in range(3)]
-
-        APT = np.zeros((natom, 3, 3))
-        for A in range(natom):
-            dip = self.ci.derivatives.dipole(A)
-            for beta in range(3):
-                pert = Perturbation('nuclear', (A, beta))
-                U_R = np.asarray(self.ci.cphf.full_U(pert))
-                dc1, dc2, dc0v = self.ci._solve_cpci(pert)
-                dc1, dc2 = dc1.real, dc2.real
-                dc0v = dc0v.real if hasattr(dc0v, 'real') else dc0v
-                dz, dD_corr = self._solve_dz_dR(pert, dc1, dc2, dc0v)
-
-                for alpha in range(3):
-                    h_RE = np.asarray(dip[alpha * 3 + beta])
-                    T1 = c('ij,ij->', Qt, h_RE)
-                    T2 = (c('rp,pq,rq->', U_R, Qt, h_E[alpha]) + c('pq,ps,sq->', Qt, h_E[alpha], U_R))
-                    T3 = c('pq,pq->', dD_corr, h_E[alpha])
-                    T3z = 2.0 * c('ai,ai->', dz, h_E[alpha][v, o])
-                    APT[A, beta, alpha] = (T1 + T2 + T3 + T3z).real
-        return APT
-
-    def _solve_dz_dR(self, pert, dc1, dc2, dc0v):
-        c = self.contract
-        o, v, no, nv = self.ci.o, self.ci.v, self.ci.no, self.ci.nv
-        zv = self.ci._zvector()
-        z = zv['z_ai']
-
-        dF, dERI, _ = self.ci._cpci_ints(pert)
-        dD_corr = self.ci._perturbed_cisd_corr_opdm(dc1, dc2)
-        dD_pqrs = self.ci._perturbed_cisd_tpdm(dc1, dc2, dc0v)
-
-        o_ = self.ci.o
-        dh_mo = dF - c('piqi->pq', 2.0 * dERI[:, o_, :, o_] - dERI.swapaxes(2, 3)[:, o_, :, o_])
-        dG_sym = 0.25 * (dD_pqrs + dD_pqrs.transpose(1, 0, 3, 2) + dD_pqrs.transpose(2, 3, 0, 1) + dD_pqrs.transpose(3, 2, 1, 0))
-        dG = dG_sym.copy()
-        dG_cross = np.zeros_like(dG)
-        for i in range(no):
-            dG_cross[i, :, i, :] += 2.0 * dD_corr
-            dG_cross[:, i, :, i] += 2.0 * dD_corr
-            dG_cross[i, :, :, i] -= dD_corr.T
-            dG_cross[:, i, i, :] -= dD_corr
-        dG = dG + 0.5 * dG_cross
-
+        t = slice(0, nbf)
+        no, nv = self.ci.no, self.ci.nv
         ERI = np.asarray(self.ci.H.ERI)
-        dX = c('jm,im->ij', dD_corr, zv['h_mo']) + c('jm,im->ij', zv['Q'], dh_mo)
-        dX = dX + 2.0 * c('jmkl,imkl->ij', dG, ERI) + 2.0 * c('jmkl,imkl->ij', zv['G'], dERI)
-
-        db = -(dX[v, o] - dX[o, v].T)
-
-        dL = 2.0 * dERI - dERI.swapaxes(2, 3)
-        dG_mat_vovo = (dL[v, o, o, v].transpose(0, 2, 3, 1)
-                       + dL[v, v, o, o].transpose(0, 2, 1, 3))
-        dG_mat_vovo = dG_mat_vovo + c('ab,ij->aibj', dF[v, v], np.eye(no))
-        dG_mat_vovo = dG_mat_vovo - c('ab,ij->aibj', np.eye(nv), dF[o, o])
-        dGz = c('aibj,bj->ai', dG_mat_vovo, z)
-
-        rhs = db - dGz
-        dz_ov = self.ci.cphf.solve(rhs.T)
-        dz = dz_ov.T
-        return dz, dD_corr
-
-
-    # Hessian
-
-    def _hess_build_Y(self):
-        """Y_ijkl double-UU tensor, correlation densities only (linear in Q/G,
-        so the HF part belongs to the SCF reference Hessian)."""
-        ct = self.contract
-        zv = self.ci._zvector()
-        Q, G, _ = self.ci._corr_QGX()
-        h_mo = zv['h_mo']
-        ERI = np.asarray(self.ci.H.ERI)
-        Y = ct('jl,ik->ijkl', Q, h_mo)
-        Y = Y + 2.0 * ct('jlmn,ikmn->ijkl', G, ERI)
-        Y = Y + 2.0 * ct('jmln,imkn->ijkl', G, ERI)
-        Y = Y + 2.0 * ct('jmnl,imnk->ijkl', G, ERI)
-        return Y
-
-    def _hess_X_tilde(self):
-        """Correlation-only X_tilde: X_corr (linear in the corr densities) with
-        the Z-vector-corrected vo block. z itself is solved from the FULL
-        Lagrangian in _zvector() (unchanged from the validated code); it is a
-        pure correlation quantity since the HF part of X satisfies Brillouin
-        (X_HF[v,o] - X_HF[o,v].T = 0) and contributes nothing to the rhs."""
-        ct = self.contract
-        o, v, no, nv = self.ci.o, self.ci.v, self.ci.no, self.ci.nv
-        zv = self.ci._zvector()
-        _, _, X_corr = self.ci._corr_QGX()
-        F, ERI = np.asarray(self.ci.H.F), np.asarray(self.ci.H.ERI)
-        A = (2.0 * ERI - ERI.swapaxes(2, 3)) + (2.0 * ERI - ERI.swapaxes(2, 3)).swapaxes(1, 3)
-        A = A.swapaxes(1, 2)
-        G_mat = (ct('ab,ij->aibj', np.eye(nv), np.eye(no)) * (F[v, v].reshape(nv, 1, nv, 1) - F[o, o].reshape(1, no, 1, no)) + A[v, o, v, o])
-        X_tilde = X_corr.copy()
-        X_tilde[v, o] += ct('aibj,bj->ai', G_mat, zv['z_ai'])
-        return X_tilde
-
-    def build_xi_ab(self, U_a, U_b, S_a, S_b, S_ab):
-        c = self.contract
-        xi = S_ab.copy()
-        xi = xi + c('im,jm->ij', U_a, U_b) + c('im,jm->ij', U_b, U_a)
-        xi = xi - c('im,jm->ij', S_a, S_b) - c('im,jm->ij', S_b, S_a)
-        return xi
-
-    def build_X_deriv(self, h_a_skel, g_a_skel):
-        """Skeleton-derivative Lagrangian, correlation densities only."""
-        ct = self.contract
-        Q, G, _ = self.ci._corr_QGX()
-        Xa = ct('jm,im->ij', Q, h_a_skel)
-        Xa = Xa + 2.0 * ct('jmkl,imkl->ij', G, g_a_skel)
-        return Xa
-
-    def build_cpci_term(self, dc1_a, dc2_a, dc1_b, dc2_b):
-        """The pure linear CI-Jacobian action on dc1_b/dc2_b."""
-        c = self.contract
-        o, v = self.ci.o, self.ci.v
-        F, ERI = np.asarray(self.ci.H.F), np.asarray(self.ci.H.ERI)
-        E_tot = self.ci.eci
-
-        sig1 = -E_tot * dc1_b
-        sig1 = sig1 - c('ji,ja->ia', F[o, o], dc1_b)
-        sig1 = sig1 + c('ab,ib->ia', F[v, v], dc1_b)
-        sig1 = sig1 + c('jabi,jb->ia', 2.0 * ERI[o, v, v, o] - ERI.swapaxes(2, 3)[o, v, v, o], dc1_b)
-        sig1 = sig1 + c('jb,ijab->ia', F[o, v], 2.0 * dc2_b - dc2_b.swapaxes(2, 3))
-        sig1 = sig1 + c('ajbc,ijbc->ia', 2.0 * ERI[v, o, v, v] - ERI.swapaxes(2, 3)[v, o, v, v], dc2_b)
-        sig1 = sig1 - c('kjib,kjab->ia', 2.0 * ERI[o, o, o, v] - ERI.swapaxes(2, 3)[o, o, o, v], dc2_b)
-
-        sig2 = -E_tot * dc2_b
-        sig2 = sig2 + c('abcj,ic->ijab', ERI[v, v, v, o], dc1_b)
-        sig2 = sig2 + c('abic,jc->ijab', ERI[v, v, o, v], dc1_b)
-        sig2 = sig2 - c('kbij,ka->ijab', ERI[o, v, o, o], dc1_b)
-        sig2 = sig2 - c('akij,kb->ijab', ERI[v, o, o, o], dc1_b)
-        sig2 = sig2 + c('ac,ijcb->ijab', F[v, v], dc2_b)
-        sig2 = sig2 + c('bc,ijac->ijab', F[v, v], dc2_b)
-        sig2 = sig2 - c('ki,kjab->ijab', F[o, o], dc2_b)
-        sig2 = sig2 - c('kj,ikab->ijab', F[o, o], dc2_b)
-        sig2 = sig2 + c('klij,klab->ijab', ERI[o, o, o, o], dc2_b)
-        sig2 = sig2 + c('abcd,ijcd->ijab', ERI[v, v, v, v], dc2_b)
-        sig2 = sig2 - c('kbcj,ikca->ijab', ERI[o, v, v, o], dc2_b)
-        sig2 = sig2 + c('kaci,kjcb->ijab', 2.0 * ERI[o, v, v, o] - ERI.swapaxes(2, 3)[o, v, v, o], dc2_b)
-        sig2 = sig2 - c('kbic,kjac->ijab', ERI[o, v, o, v], dc2_b)
-        sig2 = sig2 - c('kaci,kjbc->ijab', ERI[o, v, v, o], dc2_b)
-        sig2 = sig2 + c('kbcj,ikac->ijab', 2.0 * ERI[o, v, v, o] - ERI.swapaxes(2, 3)[o, v, v, o], dc2_b)
-        sig2 = sig2 - c('kajc,ikcb->ijab', ERI[o, v, o, v], dc2_b)
-
-        return -2.0 * (2.0 * c('ia,ia->', dc1_a, sig1) + c('ijab,ijab->', 2.0 * dc2_a - dc2_a.swapaxes(2, 3), sig2)).real
-
-    def hessian(self, route: str = 'explicit') -> np.ndarray:
-        """CISD CORRELATION-ONLY nuclear Hessian, shape (3*natom, 3*natom):
-        the T0-T5 assembly with the correlation densities in the density with linear
-        terms (T0, xi-X_tilde, T3, T4, the HF blocks of those terms are the
-        SCF reference Hessian) and the unchanged pure-correlation Z-vector (Bab.z)
-        and CPCI (T5) terms; nuclear-repulsion block dropped. Reference and
-        nuclear are supplied by the pycc.hessian facade, matching MPwfn's
-        contract."""
-        from .cphf import Perturbation
-        ct = self.contract
-        o, v, no = self.ci.o, self.ci.v, self.ci.no
-        mints, C_p4 = self.ci._psi4_mints()
-        mol = self.ci.ref.molecule()
-        natom = mol.natom()
-        ndof = 3 * natom
         F = np.asarray(self.ci.H.F)
+        C = np.asarray(self.ci.C)
         eps = np.diag(F)
+        mints, _ = self._psi4_mints()
 
-        zv = self.ci._zvector()
-        Q, G, _ = self.ci._corr_QGX()   # correlation densities (T0/T3/T4/xi-X)
-        X_t = self._hess_X_tilde()
-        Y = self._hess_build_Y()
+        # A-matrix (antisymmetric perturbation)
+        A_mag = -(2 * ERI - ERI.swapaxes(2, 3)) + (2 * ERI - ERI.swapaxes(2, 3)).swapaxes(1, 3)
+        A_mag = A_mag.swapaxes(1, 2)
+        G_mag = (ct('ab,ij,aibj->aibj', np.eye(nv), np.eye(no), F[v, v].reshape(nv, 1, nv, 1) - F[o, o].reshape(1, no, 1, no)) + A_mag[v, o, v, o])
+        G_mag = np.linalg.inv(G_mag.reshape(nv * no, nv * no))
+
+        # Magnetic dipole integrals
+        L_AO = mints.ao_angular_momentum()
+        h_mag = ct('mp,mn,nq->pq', C.conj(), -0.5 * L_AO[beta].np, C)
+
+        # U_H  (full 4-block solve - exactly as in working code)
+        U_H = np.zeros((nbf, nbf), dtype=complex)
+        B_vo = h_mag[v, o]
+        U_H[v, o] += (G_mag @ B_vo.reshape(nv * no)).reshape(nv, no)
+        U_H[o, v] += U_H[v, o].T
+
+        D_oo = (eps[o] - eps[o].reshape(-1, 1)) + np.eye(no)
+        B_oo = (-h_mag[o, o].copy() + ct('em,iejm->ij', U_H[v, o], A_mag.swapaxes(1, 2)[o, v, o, o]))
+        U_H[o, o] += B_oo / D_oo
+
+        D_vv = (eps[v] - eps[v].reshape(-1, 1)) + np.eye(nv)
+        B_vv = (-h_mag[v, v].copy() + ct('em,aebm->ab', U_H[v, o], A_mag.swapaxes(1, 2)[v, v, v, o]))
+        U_H[v, v] += B_vv / D_vv
+
+        for j in range(no):
+            U_H[j, j] = 0
+        for cc in range(no, nbf):
+            U_H[cc, cc] = 0
+
+        # dF/dH
+        dF = np.zeros((nbf, nbf), dtype=complex)
+        dF[o, o] -= h_mag[o, o].copy()
+        dF[o, o] += (U_H[o, o] * eps[o].reshape(-1, 1) - U_H[o, o].swapaxes(0, 1) * eps[o])
+        dF[o, o] += ct('em,iejm->ij', U_H[v, o], A_mag.swapaxes(1, 2)[o, v, o, o])
+
+        dF[v, v] -= h_mag[v, v].copy()
+        dF[v, v] += (U_H[v, v] * eps[v].reshape(-1, 1) - U_H[v, v].swapaxes(0, 1) * eps[v])
+        dF[v, v] += ct('em,aebm->ab', U_H[v, o], A_mag.swapaxes(1, 2)[v, v, v, o])
+
+        # dERI/dH
+        dERI = np.zeros(ERI.shape, dtype=complex)
+        dERI += ct('tr,pqts->pqrs', U_H[:, t], ERI[t, t, :, t])
+        dERI += ct('ts,pqrt->pqrs', U_H[:, t], ERI[t, t, t, :])
+        dERI -= ct('tp,tqrs->pqrs', U_H[:, t], ERI[:, t, t, t])
+        dERI -= ct('tq,ptrs->pqrs', U_H[:, t], ERI[t, :, t, t])
+
+        return dF, dERI, U_H
+
+    def _build_vecpot_ints(self, gamma):
+        ct = self.contract
+        nbf = self.ci.nmo
+        o, v = self.ci.o, self.ci.v
+        t = slice(0, nbf)
+        no, nv = self.ci.no, self.ci.nv
         ERI = np.asarray(self.ci.H.ERI)
-        z = zv['z_ai']
-        n0 = self.ci._normalized_amplitudes()[0]
+        F = np.asarray(self.ci.H.F)
+        C = np.asarray(self.ci.C)
+        eps = np.diag(F)
+        mints, _ = self._psi4_mints()
 
-        ERI_M = ERI.transpose(0, 2, 1, 3)
-        W_M = 2 * ERI_M - ERI_M.transpose(0, 2, 1, 3)
-        A_M = 4 * ERI_M - ERI_M.transpose(0, 2, 1, 3) - ERI_M.transpose(0, 3, 2, 1)
+        # A-matrix for antisymmetric perturbation (same as magnetic)
+        A_mag = -(2 * ERI - ERI.swapaxes(2, 3)) + (2 * ERI - ERI.swapaxes(2, 3)).swapaxes(1, 3)
+        A_mag = A_mag.swapaxes(1, 2)
+        G_mag = (ct('ab,ij,aibj->aibj', np.eye(nv), np.eye(no), F[v, v].reshape(nv, 1, nv, 1) - F[o, o].reshape(1, no, 1, no)) + A_mag[v, o, v, o])
+        G_mag = np.linalg.inv(G_mag.reshape(nv * no, nv * no))
 
-        def to_mulliken(g_dirac):
-            return g_dirac.transpose(0, 2, 1, 3)
+        # Linear momentum integrals: p_gamma = -i * nabla_gamma
+        nabla_AO = mints.ao_nabla()
+        h_A = ct('mp,mn,nq->pq', C.conj(), -nabla_AO[gamma].np, C)
 
-        h_skel_1, g_skel_1, S_skel_1 = {}, {}, {}
-        U_all, dT1_all, dT2_all = {}, {}, {}
+        # Solve CPHF for U^A (identical structure to U^H)
+        U_A = np.zeros((nbf, nbf), dtype=complex)
+        B_vo = h_A[v, o]
+        U_A[v, o] += (G_mag @ B_vo.reshape(nv * no)).reshape(nv, no)
+        U_A[o, v] += U_A[v, o].T
 
-        for N1 in range(natom):
-            T_core = mints.mo_oei_deriv1('KINETIC', N1, C_p4, C_p4)
-            V_core = mints.mo_oei_deriv1('POTENTIAL', N1, C_p4, C_p4)
-            S_core = mints.mo_oei_deriv1('OVERLAP', N1, C_p4, C_p4)
-            E_core = mints.mo_tei_deriv1(N1, C_p4, C_p4, C_p4, C_p4)
-            for a in range(3):
-                la = 3 * N1 + a
-                h_skel_1[la] = T_core[a].np + V_core[a].np
-                S_skel_1[la] = S_core[a].np
-                g_skel_1[la] = E_core[a].np.swapaxes(1, 2)
-                pert = Perturbation('nuclear', (N1, a))
-                U_all[la] = np.asarray(self.ci.cphf.full_U(pert))
-                dT1_all[la], dT2_all[la] = self.ci._cpci_raw(pert)
+        D_oo = (eps[o] - eps[o].reshape(-1, 1)) + np.eye(no)
+        B_oo = (-h_A[o, o].copy() + ct('em,iejm->ij', U_A[v, o], A_mag.swapaxes(1, 2)[o, v, o, o]))
+        U_A[o, o] += B_oo / D_oo
 
-        X_deriv = {la: self.build_X_deriv(h_skel_1[la], g_skel_1[la]) for la in range(ndof)}
+        D_vv = (eps[v] - eps[v].reshape(-1, 1)) + np.eye(nv)
+        B_vv = (-h_A[v, v].copy() + ct('em,aebm->ab', U_A[v, o], A_mag.swapaxes(1, 2)[v, v, v, o]))
+        U_A[v, v] += B_vv / D_vv
 
-        Fa_M, Aa_M = {}, {}
-        for la in range(ndof):
-            g_a_M = to_mulliken(g_skel_1[la])
-            Fa = h_skel_1[la].copy()
-            Fa += (ct('ijkk->ij', 2 * g_a_M[:, :, o, o]) - ct('ikjk->ij', g_a_M[:, o, :, o]))
-            Fa_M[la] = Fa
-            Aa_M[la] = (4 * g_a_M - g_a_M.transpose(0, 2, 1, 3) - g_a_M.transpose(0, 3, 2, 1))
+        for j in range(no):
+            U_A[j, j] = 0
+        for cc in range(no, nbf):
+            U_A[cc, cc] = 0
 
-        Hessian = np.zeros((ndof, ndof))
+        # dF/dA_gamma
+        dF = np.zeros((nbf, nbf), dtype=complex)
+        dF[o, o] -= h_A[o, o].copy()
+        dF[o, o] += (U_A[o, o] * eps[o].reshape(-1, 1) - U_A[o, o].swapaxes(0, 1) * eps[o])
+        dF[o, o] += ct('em,iejm->ij', U_A[v, o], A_mag.swapaxes(1, 2)[o, v, o, o])
 
-        def _assemble_ab(a, b, h_ab_sk, S_ab_sk, g_ab_sk, N1, N2):
-            la, lb = 3 * N1 + a, 3 * N2 + b
-            U_a, U_b = U_all[la], U_all[lb]
+        dF[v, v] -= h_A[v, v].copy()
+        dF[v, v] += (U_A[v, v] * eps[v].reshape(-1, 1) - U_A[v, v].swapaxes(0, 1) * eps[v])
+        dF[v, v] += ct('em,aebm->ab', U_A[v, o], A_mag.swapaxes(1, 2)[v, v, v, o])
 
-            # T1: skeleton density contraction
-            t0_val = ct('ij,ij->', Q, h_ab_sk)
-            t0_val += ct('pqrs,pqrs->', G, g_ab_sk)
+        # dERI/dA_gamma (orbital response only, same sign pattern as magnetic)
+        dERI = np.zeros(ERI.shape, dtype=complex)
+        dERI += ct('tr,pqts->pqrs', U_A[:, t], ERI[t, t, :, t])
+        dERI += ct('ts,pqrt->pqrs', U_A[:, t], ERI[t, t, t, :])
+        dERI -= ct('tp,tqrs->pqrs', U_A[:, t], ERI[:, t, t, t])
+        dERI -= ct('tq,ptrs->pqrs', U_A[:, t], ERI[t, :, t, t])
 
-            # T2: Z-vector replacement
-            xi = self.build_xi_ab(U_a, U_b, S_skel_1[la], S_skel_1[lb], S_ab_sk)
-            t2_xiXt = -ct('ij,ji->', xi, X_t)
-            g_ab_sk_M = to_mulliken(g_ab_sk)
-            F_ab_M = (h_ab_sk + ct('ijkk->ij', 2 * g_ab_sk_M[:, :, o, o]) - ct('ikjk->ij', g_ab_sk_M[:, o, :, o]))
-            Bab = F_ab_M.copy()
-            Bab -= ct('ij,j->ij', xi, eps)
-            Bab -= ct('kl,ijkl->ij', xi[o, o], W_M[:, :, o, o])
-            Bab += ct('ki,kj->ij', U_a, Fa_M[lb])
-            Bab += ct('ki,kj->ij', U_b, Fa_M[la])
-            Bab += ct('kj,ik->ij', U_a, Fa_M[lb])
-            Bab += ct('kj,ik->ij', U_b, Fa_M[la])
-            Bab += ct('ki,kj,k->ij', U_a, U_b, eps)
-            Bab += ct('ki,kj,k->ij', U_b, U_a, eps)
-            UaUbT = U_a[:, o] @ U_b[:, o].T
-            Bab += ct('kl,ijkl->ij', UaUbT, A_M)
-            temp5_b = ct('lm,kjlm->kj', U_b[:, o], A_M[:, :, :, o])
-            temp5_a = ct('lm,kjlm->kj', U_a[:, o], A_M[:, :, :, o])
-            Bab += ct('ki,kj->ij', U_a, temp5_b)
-            Bab += ct('ki,kj->ij', U_b, temp5_a)
-            temp6_b = ct('lm,iklm->ik', U_b[:, o], A_M[:, :, :, o])
-            temp6_a = ct('lm,iklm->ik', U_a[:, o], A_M[:, :, :, o])
-            Bab += ct('kj,ik->ij', U_a, temp6_b)
-            Bab += ct('kj,ik->ij', U_b, temp6_a)
-            Bab += ct('kl,ijkl->ij', U_a[:, o], Aa_M[lb][:, :, :, o])
-            Bab += ct('kl,ijkl->ij', U_b[:, o], Aa_M[la][:, :, :, o])
-            t2_zvec = 2.0 * ct('ai,ai->', Bab[v, o], z)
-            t2_val = t2_xiXt + t2_zvec
+        return dF, dERI, U_A
 
-            # T3+T4: orbital-response Lagrangian + Y tensor
-            t3_val = 2.0 * (ct('ij,ij->', U_b, X_deriv[la])+ ct('ij,ij->', U_a, X_deriv[lb]))
-            t4_val = 2.0 * ct('ij,kl,ijkl->', U_a, U_b, Y)
+    def _cpci_ints(self, pert):
+        """(dF, dERI, U) for a cphf.Perturbation. Used for magnetic/vecpot (AAT/VG-APT) and for
+        CIwfn's own (non-canonical, active-space) nuclear response - NOT for the field/nuclear
+        perturbations that feed the base's second-derivative machinery, which instead thread the
+        base's own canonical, full-occupied-space (df, deri) directly into _solve_cpci_ints."""
+        if getattr(self, '_cpci_ints_cache', None) is None:
+            self._cpci_ints_cache = {}
+        if pert in self._cpci_ints_cache:
+            return self._cpci_ints_cache[pert]
+        cphf = self.ci.cphf
+        if pert.kind == 'nuclear':
+            dF = np.asarray(cphf.perturbed_fock(pert))
+            dERI = np.asarray(cphf.perturbed_eri(pert))
+            U = np.asarray(cphf.full_U(pert))
+            result = (dF, dERI, U)
+        elif pert.kind == 'magnetic':
+            result = self._build_magnetic_ints(pert.comp)
+        elif pert.kind == 'vecpot':
+            result = self._build_vecpot_ints(pert.comp)
+        elif pert.kind == 'field':
+            dF = np.asarray(cphf.perturbed_fock(pert))
+            dERI = np.asarray(cphf.perturbed_eri(pert))   # zero for electric field
+            U = np.asarray(cphf.full_U(pert))
+            result = (dF, dERI, U)
+        else:
+            raise ValueError(f"unknown perturbation kind {pert.kind!r}")
+        self._cpci_ints_cache[pert] = result
+        return result
 
-            # T5: CPCI amplitude response
-            t5_val = n0**2 * self.build_cpci_term(
-                dT1_all[la], dT2_all[la], dT1_all[lb], dT2_all[lb]).real
+    def _solve_cpci_ints(self, dF, dERI, imaginary=False, maxiter=100, diis_start=2, diis_max=8,
+                          e_convergence=1e-11, d_convergence=1e-11):
+        """Core coupled-perturbed-CI iterative solve given already-built perturbed integrals
+        (dF, dERI) - the CISD analog of CCderiv._perturbed_amplitudes, factored out so both
+        CIderiv's own _cpci_ints-driven entry point (_solve_cpci, below) and the base's
+        canonical/full-occupied-space (df, deri) (via _perturbed_unrelaxed_densities) can drive
+        the same solve. Returns (dc1, dc2, dc0v, dt1, dt2): the true-normalized response
+        (dc1, dc2, dc0v) and the raw intermediate-normalized response (dt1, dt2).
 
-            return la, lb, (t0_val + t2_val + t3_val + t4_val + t5_val).real
+        `imaginary=True` for magnetic/vecpot-type perturbations (dc0v forced to 0 - the real
+        normalization response vanishes by time-reversal symmetry for an imaginary perturbation);
+        `False` (default) for real perturbations (nuclear/field), where dc0v is generally
+        nonzero."""
+        from .utils import helper_diis
+        ci = self.ci
+        c = self.contract
+        o, v = ci.o, ci.v
+        t1, t2 = ci.c1, ci.c2
+        F, ERI = np.asarray(ci.H.F), np.asarray(ci.H.ERI)
+        Dia, Dijab = ci.Dia, ci.Dijab
+        E_cisd = ci.eci
+        n0 = ci._normalized_amplitudes()[0]
+        dF, dERI = np.asarray(dF), np.asarray(dERI)
 
-        for N1 in range(natom):
-            for N2 in range(natom):
-                T_ab = mints.mo_oei_deriv2('KINETIC', N1, N2, C_p4, C_p4)
-                V_ab = mints.mo_oei_deriv2('POTENTIAL', N1, N2, C_p4, C_p4)
-                S_ab = mints.mo_oei_deriv2('OVERLAP', N1, N2, C_p4, C_p4)
-                E_ab = mints.mo_tei_deriv2(N1, N2, C_p4, C_p4, C_p4, C_p4)
+        D_pq, D_pq_corr, D_pqrs = ci._cisd_densities()
+        dE = c('pq,pq->', dF, D_pq) + c('pqrs,pqrs->', dERI, D_pqrs)
 
-                h_ab = {(a, b): T_ab[3 * a + b].np + V_ab[3 * a + b].np for a in range(3) for b in range(3)}
-                S_ab_np = {(a, b): S_ab[3 * a + b].np for a in range(3) for b in range(3)}
-                g_ab = {(a, b): E_ab[3 * a + b].np.swapaxes(1, 2) for a in range(3) for b in range(3)}
+        dt1 = -(dE * t1).astype(complex)
+        dt1 = dt1 - c('ji,ja->ia', dF[o, o], t1)
+        dt1 = dt1 + c('ab,ib->ia', dF[v, v], t1)
+        dt1 = dt1 + c('jabi,jb->ia', 2.0 * dERI[o, v, v, o] - dERI.swapaxes(2, 3)[o, v, v, o], t1)
+        dt1 = dt1 + c('jb,ijab->ia', dF[o, v], 2.0 * t2 - t2.swapaxes(2, 3))
+        dt1 = dt1 + c('ajbc,ijbc->ia', 2.0 * dERI[v, o, v, v] - dERI.swapaxes(2, 3)[v, o, v, v], t2)
+        dt1 = dt1 - c('kjib,kjab->ia', 2.0 * dERI[o, o, o, v] - dERI.swapaxes(2, 3)[o, o, o, v], t2)
+        dt1 = dt1 / Dia
 
-                for a in range(3):
-                    for b in range(3):
-                        la, lb, val = _assemble_ab(
-                            a, b, h_ab[(a, b)], S_ab_np[(a, b)], g_ab[(a, b)], N1, N2)
-                        Hessian[la, lb] = val
+        dt2 = -(dE * t2).astype(complex)
+        dt2 = dt2 + c('abcj,ic->ijab', dERI[v, v, v, o], t1)
+        dt2 = dt2 + c('abic,jc->ijab', dERI[v, v, o, v], t1)
+        dt2 = dt2 - c('kbij,ka->ijab', dERI[o, v, o, o], t1)
+        dt2 = dt2 - c('akij,kb->ijab', dERI[v, o, o, o], t1)
+        dt2 = dt2 + c('ac,ijcb->ijab', dF[v, v], t2)
+        dt2 = dt2 + c('bc,ijac->ijab', dF[v, v], t2)
+        dt2 = dt2 - c('ki,kjab->ijab', dF[o, o], t2)
+        dt2 = dt2 - c('kj,ikab->ijab', dF[o, o], t2)
+        dt2 = dt2 + c('klij,klab->ijab', dERI[o, o, o, o], t2)
+        dt2 = dt2 + c('abcd,ijcd->ijab', dERI[v, v, v, v], t2)
+        dt2 = dt2 - c('kbcj,ikca->ijab', dERI[o, v, v, o], t2)
+        dt2 = dt2 + c('kaci,kjcb->ijab', 2.0 * dERI[o, v, v, o] - dERI.swapaxes(2, 3)[o, v, v, o], t2)
+        dt2 = dt2 - c('kbic,kjac->ijab', dERI[o, v, o, v], t2)
+        dt2 = dt2 - c('kaci,kjbc->ijab', dERI[o, v, v, o], t2)
+        dt2 = dt2 + c('kbcj,ikac->ijab', 2.0 * dERI[o, v, v, o] - dERI.swapaxes(2, 3)[o, v, v, o], t2)
+        dt2 = dt2 - c('kajc,ikcb->ijab', dERI[o, v, o, v], t2)
+        dt2 = dt2 / Dijab
 
-                del T_ab, V_ab, S_ab, E_ab, h_ab, S_ab_np, g_ab
+        dE_proj = (2.0 * c('ia,ia->', t1, dF[o, v])
+                   + c('ijab,ijab->', t2, 2.0 * dERI[o, o, v, v] - dERI.swapaxes(2, 3)[o, o, v, v])
+                   + 2.0 * c('ia,ia->', dt1, F[o, v])
+                   + c('ijab,ijab->', dt2, 2.0 * ERI[o, o, v, v] - ERI.swapaxes(2, 3)[o, o, v, v]))
 
-        Hessian = 0.5 * (Hessian + Hessian.T)
-        return Hessian
+        diis = helper_diis(dt1, dt2, diis_max, getattr(ci, 'precision', 1e-12))
 
-    # CISD-specific properties (overlap formulations; not part of the base)
-    # AAT and the velocity-gauge APT are wave-function-overlap constructions (Krishnan, Shumberger &
-    # Crawford)
+        for iteration in range(1, maxiter + 1):
+            dE_proj_old = dE_proj
+            dt1_old, dt2_old = dt1.copy(), dt2.copy()
+
+            # singles residual - driving terms (dF/dERI acting on t1/t2)
+            dRt1 = dF.copy().swapaxes(0, 1)[o, v].astype(complex)
+            dRt1 = dRt1 - dE_proj * t1
+            dRt1 = dRt1 - c('ji,ja->ia', dF[o, o], t1)
+            dRt1 = dRt1 + c('ab,ib->ia', dF[v, v], t1)
+            dRt1 = dRt1 + c('jabi,jb->ia', 2.0 * dERI[o, v, v, o] - dERI.swapaxes(2, 3)[o, v, v, o], t1)
+            dRt1 = dRt1 + c('jb,ijab->ia', dF[o, v], 2.0 * t2 - t2.swapaxes(2, 3))
+            dRt1 = dRt1 + c('ajbc,ijbc->ia', 2.0 * dERI[v, o, v, v] - dERI.swapaxes(2, 3)[v, o, v, v], t2)
+            dRt1 = dRt1 - c('kjib,kjab->ia', 2.0 * dERI[o, o, o, v] - dERI.swapaxes(2, 3)[o, o, o, v], t2)
+            # singles residual - response terms (F/ERI acting on dt1/dt2)
+            dRt1 = dRt1 - E_cisd * dt1
+            dRt1 = dRt1 - c('ji,ja->ia', F[o, o], dt1)
+            dRt1 = dRt1 + c('ab,ib->ia', F[v, v], dt1)
+            dRt1 = dRt1 + c('jabi,jb->ia', 2.0 * ERI[o, v, v, o] - ERI.swapaxes(2, 3)[o, v, v, o], dt1)
+            dRt1 = dRt1 + c('jb,ijab->ia', F[o, v], 2.0 * dt2 - dt2.swapaxes(2, 3))
+            dRt1 = dRt1 + c('ajbc,ijbc->ia', 2.0 * ERI[v, o, v, v] - ERI.swapaxes(2, 3)[v, o, v, v], dt2)
+            dRt1 = dRt1 - c('kjib,kjab->ia', 2.0 * ERI[o, o, o, v] - ERI.swapaxes(2, 3)[o, o, o, v], dt2)
+
+            # doubles residual - driving terms
+            dRt2 = dERI.copy().swapaxes(0, 2).swapaxes(1, 3)[o, o, v, v].astype(complex)
+            dRt2 = dRt2 - dE_proj * t2
+            dRt2 = dRt2 + c('abcj,ic->ijab', dERI[v, v, v, o], t1)
+            dRt2 = dRt2 + c('abic,jc->ijab', dERI[v, v, o, v], t1)
+            dRt2 = dRt2 - c('kbij,ka->ijab', dERI[o, v, o, o], t1)
+            dRt2 = dRt2 - c('akij,kb->ijab', dERI[v, o, o, o], t1)
+            dRt2 = dRt2 + c('ac,ijcb->ijab', dF[v, v], t2)
+            dRt2 = dRt2 + c('bc,ijac->ijab', dF[v, v], t2)
+            dRt2 = dRt2 - c('ki,kjab->ijab', dF[o, o], t2)
+            dRt2 = dRt2 - c('kj,ikab->ijab', dF[o, o], t2)
+            dRt2 = dRt2 + c('klij,klab->ijab', dERI[o, o, o, o], t2)
+            dRt2 = dRt2 + c('abcd,ijcd->ijab', dERI[v, v, v, v], t2)
+            dRt2 = dRt2 - c('kbcj,ikca->ijab', dERI[o, v, v, o], t2)
+            dRt2 = dRt2 + c('kaci,kjcb->ijab', 2.0 * dERI[o, v, v, o] - dERI.swapaxes(2, 3)[o, v, v, o], t2)
+            dRt2 = dRt2 - c('kbic,kjac->ijab', dERI[o, v, o, v], t2)
+            dRt2 = dRt2 - c('kaci,kjbc->ijab', dERI[o, v, v, o], t2)
+            dRt2 = dRt2 + c('kbcj,ikac->ijab', 2.0 * dERI[o, v, v, o] - dERI.swapaxes(2, 3)[o, v, v, o], t2)
+            dRt2 = dRt2 - c('kajc,ikcb->ijab', dERI[o, v, o, v], t2)
+            # doubles residual - response terms
+            dRt2 = dRt2 - E_cisd * dt2
+            dRt2 = dRt2 + c('abcj,ic->ijab', ERI[v, v, v, o], dt1)
+            dRt2 = dRt2 + c('abic,jc->ijab', ERI[v, v, o, v], dt1)
+            dRt2 = dRt2 - c('kbij,ka->ijab', ERI[o, v, o, o], dt1)
+            dRt2 = dRt2 - c('akij,kb->ijab', ERI[v, o, o, o], dt1)
+            dRt2 = dRt2 + c('ac,ijcb->ijab', F[v, v], dt2)
+            dRt2 = dRt2 + c('bc,ijac->ijab', F[v, v], dt2)
+            dRt2 = dRt2 - c('ki,kjab->ijab', F[o, o], dt2)
+            dRt2 = dRt2 - c('kj,ikab->ijab', F[o, o], dt2)
+            dRt2 = dRt2 + c('klij,klab->ijab', ERI[o, o, o, o], dt2)
+            dRt2 = dRt2 + c('abcd,ijcd->ijab', ERI[v, v, v, v], dt2)
+            dRt2 = dRt2 - c('kbcj,ikca->ijab', ERI[o, v, v, o], dt2)
+            dRt2 = dRt2 + c('kaci,kjcb->ijab', 2.0 * ERI[o, v, v, o] - ERI.swapaxes(2, 3)[o, v, v, o], dt2)
+            dRt2 = dRt2 - c('kbic,kjac->ijab', ERI[o, v, o, v], dt2)
+            dRt2 = dRt2 - c('kaci,kjbc->ijab', ERI[o, v, v, o], dt2)
+            dRt2 = dRt2 + c('kbcj,ikac->ijab', 2.0 * ERI[o, v, v, o] - ERI.swapaxes(2, 3)[o, v, v, o], dt2)
+            dRt2 = dRt2 - c('kajc,ikcb->ijab', ERI[o, v, o, v], dt2)
+
+            dt1 = dt1 + dRt1 / Dia
+            dt2 = dt2 + dRt2 / Dijab
+
+            diis.add_error_vector(dt1, dt2)
+            if iteration >= diis_start:
+                dt1, dt2 = diis.extrapolate(dt1, dt2)
+
+            dE_proj = (2.0 * c('ia,ia->', t1, dF[o, v])
+                       + c('ijab,ijab->', t2, 2.0 * dERI[o, o, v, v] - dERI.swapaxes(2, 3)[o, o, v, v])
+                       + 2.0 * c('ia,ia->', dt1, F[o, v])
+                       + c('ijab,ijab->', dt2, 2.0 * ERI[o, o, v, v] - ERI.swapaxes(2, 3)[o, o, v, v]))
+
+            delta_dE = abs(dE_proj - dE_proj_old)
+            rms_dt1 = np.sqrt(np.sum((dt1 - dt1_old) ** 2))
+            rms_dt2 = np.sqrt(np.sum((dt2 - dt2_old) ** 2))
+            if iteration > 1 and (delta_dE < e_convergence and rms_dt1 < d_convergence
+                                   and rms_dt2 < d_convergence):
+                break
+
+        dc0 = ci._cisd_dn0(dt1, dt2)
+        if imaginary:
+            dc0v = 0.0
+            dc1 = n0 * dt1
+            dc2 = n0 * dt2
+        else:
+            dc0v = dc0
+            dc1 = dc0 * t1 + n0 * dt1
+            dc2 = dc0 * t2 + n0 * dt2
+
+        return dc1, dc2, dc0v, dt1, dt2
+
+    def _solve_cpci(self, pert, maxiter=100, diis_start=2, diis_max=8,
+                     e_convergence=1e-11, d_convergence=1e-11):
+        """Coupled-perturbed CI, entry point keyed by Perturbation - used by the AAT/VG-APT
+        overlap code below (magnetic/vecpot/nuclear via _cpci_ints). Cached by pert. Returns
+        (dc1, dc2, dc0v). The base's second-derivative machinery does NOT go through this -
+        see _perturbed_unrelaxed_densities, which calls _solve_cpci_ints directly with the
+        base's own canonical (df, deri)."""
+        if getattr(self, '_cpci_cache', None) is None:
+            self._cpci_cache = {}
+        if pert in self._cpci_cache:
+            return self._cpci_cache[pert]
+        dF, dERI, U = self._cpci_ints(pert)
+        imaginary = pert.kind in ('magnetic', 'vecpot')
+        dc1, dc2, dc0v, dt1, dt2 = self._solve_cpci_ints(
+            dF, dERI, imaginary=imaginary, maxiter=maxiter, diis_start=diis_start,
+            diis_max=diis_max, e_convergence=e_convergence, d_convergence=d_convergence)
+        if getattr(self, '_cpci_raw_cache', None) is None:
+            self._cpci_raw_cache = {}
+        self._cpci_raw_cache[pert] = (dt1, dt2)
+        result = (dc1, dc2, dc0v)
+        self._cpci_cache[pert] = result
+        return result
+
+    def _cpci_raw(self, pert):
+        if getattr(self, '_cpci_raw_cache', None) is None or pert not in self._cpci_raw_cache:
+            self._solve_cpci(pert)
+        return self._cpci_raw_cache[pert]
+
+    # raw perturbed correlation-density builders (true-normalized)
+
+    def _perturbed_cisd_corr_opdm(self, dc1, dc2):
+        ci = self.ci
+        c = self.contract
+        o, v, nmo = ci.o, ci.v, ci.nmo
+        n0, n1, n2, tau_n = ci._normalized_amplitudes()
+        dtau_n = 2.0 * dc2 - dc2.swapaxes(2, 3)
+        sigma = n2 - n2.swapaxes(2, 3)
+        dsigma = dc2 - dc2.swapaxes(2, 3)
+
+        dD = np.zeros((nmo, nmo), dtype=dc1.dtype)
+        dD[o, o] -= 2.0 * c('ja,ia->ij', dc1, n1) + 2.0 * c('ja,ia->ij', n1, dc1)
+        dD[o, o] -= 2.0 * c('jkab,ikab->ij', dtau_n, n2) + 2.0 * c('jkab,ikab->ij', tau_n, dc2)
+        dD[v, v] += 2.0 * c('ia,ib->ab', dc1, n1) + 2.0 * c('ia,ib->ab', n1, dc1)
+        dD[v, v] += 2.0 * c('ijac,ijbc->ab', dtau_n, n2) + 2.0 * c('ijac,ijbc->ab', tau_n, dc2)
+        dD[o, v] += 2.0 * dc1
+        dD[o, v] += 2.0 * c('jb,ijab->ia', dc1, sigma) + 2.0 * c('jb,ijab->ia', n1, dsigma)
+        dD[v, o] = dD[o, v].T
+        return dD
+
+    def _perturbed_cisd_tpdm(self, dc1, dc2, dc0v):
+        ci = self.ci
+        c = self.contract
+        o, v, nmo = ci.o, ci.v, ci.nmo
+        n0, n1, n2, tau_n = ci._normalized_amplitudes()
+        dtau_n = 2.0 * dc2 - dc2.swapaxes(2, 3)
+
+        dG = np.zeros((nmo, nmo, nmo, nmo), dtype=dc1.dtype)
+        dG[o, o, o, o] = c('klab,ijab->ijkl', dc2, tau_n) + c('klab,ijab->ijkl', n2, dtau_n)
+        dG[v, v, v, v] = c('ijab,ijcd->abcd', dc2, tau_n) + c('ijab,ijcd->abcd', n2, dtau_n)
+        dG[o, v, v, o] = 4.0 * (c('ja,ib->iabj', dc1, n1) + c('ja,ib->iabj', n1, dc1))
+        dG[o, v, o, v] = -2.0 * (c('ja,ib->iajb', dc1, n1) + c('ja,ib->iajb', n1, dc1))
+        dG[v, o, o, v] = 2.0 * (c('jkac,ikbc->aijb', dtau_n, tau_n) + c('jkac,ikbc->aijb', tau_n, dtau_n))
+        dG[v, o, v, o] = (
+            -4.0 * (c('jkac,ikbc->aibj', dc2, n2) + c('jkac,ikbc->aibj', n2, dc2))
+            + 2.0 * (c('jkac,ikcb->aibj', dc2, n2) + c('jkac,ikcb->aibj', n2, dc2))
+            + 2.0 * (c('jkca,ikbc->aibj', dc2, n2) + c('jkca,ikbc->aibj', n2, dc2))
+            - 4.0 * (c('jkca,ikcb->aibj', dc2, n2) + c('jkca,ikcb->aibj', n2, dc2)))
+        dG[o, o, v, v] = dc0v * tau_n + n0 * dtau_n
+        tau_swp = 2.0 * n2.swapaxes(0, 2).swapaxes(1, 3) - n2.swapaxes(2, 3).swapaxes(0, 2).swapaxes(1, 3)
+        dtau_swp = 2.0 * dc2.swapaxes(0, 2).swapaxes(1, 3) - dc2.swapaxes(2, 3).swapaxes(0, 2).swapaxes(1, 3)
+        dG[v, v, o, o] = dc0v * tau_swp + n0 * dtau_swp
+        dG[v, o, v, v] = 2.0 * (c('ja,ijcb->aibc', dc1, tau_n) + c('ja,ijcb->aibc', n1, dtau_n))
+        dG[o, v, o, o] = -2.0 * (c('kjab,ib->iajk', dtau_n, n1) + c('kjab,ib->iajk', tau_n, dc1))
+        dG[v, v, v, o] = 2.0 * (c('jiab,jc->abci', dtau_n, n1) + c('jiab,jc->abci', tau_n, dc1))
+        dG[o, o, o, v] = -2.0 * (c('kb,ijba->ijka', dc1, tau_n) + c('kb,ijba->ijka', n1, dtau_n))
+        return dG
 
     def _build_Dtilde(self, dc1, dc2, dc0):
         c = self.contract
@@ -513,7 +479,7 @@ class CIderiv(CorrelatedDerivs):
         return R
 
     def _aat_dc_normalized(self, pert):
-        dc1, dc2, dc0v = self.ci._solve_cpci(pert)
+        dc1, dc2, dc0v = self._solve_cpci(pert)
         return dc0v, dc1, dc2
 
     def compute_Icc_AATs(self):
@@ -543,7 +509,7 @@ class CIderiv(CorrelatedDerivs):
         for la in range(3 * natom):
             A, beta_ = divmod(la, 3)
             pR = Perturbation('nuclear', (A, beta_))
-            _, _, U_R = self.ci._cpci_ints(pR)
+            _, _, U_R = self._cpci_ints(pR)
             half_S = np.asarray(self.ci.derivatives.overlap_half(A)[beta_])
             Ur_eff = U_R + half_S.T
             for beta in range(3):
@@ -561,11 +527,11 @@ class CIderiv(CorrelatedDerivs):
         for la in range(3 * natom):
             A, beta_ = divmod(la, 3)
             pR = Perturbation('nuclear', (A, beta_))
-            _, _, U_R = self.ci._cpci_ints(pR)
+            _, _, U_R = self._cpci_ints(pR)
             half_S = np.asarray(self.ci.derivatives.overlap_half(A)[beta_])
             Ur_eff = U_R + half_S.T
             for beta in range(3):
-                _, _, U_H = self.ci._cpci_ints(Perturbation('magnetic', beta))
+                _, _, U_H = self._cpci_ints(Perturbation('magnetic', beta))
                 I_pp[la, beta] = c('pq,pq->', D_pq, U_H.T @ Ur_eff).real
         return I_pp
 
@@ -580,7 +546,7 @@ class CIderiv(CorrelatedDerivs):
             dn0_R, dn1_R, dn2_R = self._aat_dc_normalized(pR)
             R_pq = self._build_Dtilde(dn1_R, dn2_R, dn0_R)
             for beta in range(3):
-                _, _, U_H = self.ci._cpci_ints(Perturbation('magnetic', beta))
+                _, _, U_H = self._cpci_ints(Perturbation('magnetic', beta))
                 AAT_cphi[la, beta] = c('pq,pq->', R_pq, U_H)
         return AAT_cphi.real
 
@@ -588,7 +554,7 @@ class CIderiv(CorrelatedDerivs):
         """CISD correlation AAT, shape (natom, 3, 3): the four electronic overlap blocks with the
         correlation 1-PDM in Iphiphi. The SCF reference and the nuclear term (Z_A/4) eps_abc R_c
         are supplied by the pycc.aat facade (HFwfn.atomic_axial_tensors + _nuclear_aat), matching
-        MPderiv's contract."""
+        MPderiv."""
         natom = self.ci.ref.molecule().natom()
         total = (self.compute_Icc_AATs() + self.compute_Iphic_AATs() + self.compute_Iphiphi_AATs() + self.compute_Icphi_AATs())
         return total.reshape(natom, 3, 3)
@@ -617,7 +583,7 @@ class CIderiv(CorrelatedDerivs):
             dn0_R, dn1_R, dn2_R = self._aat_dc_normalized(Perturbation('nuclear', (A, beta_)))
             D_tilde_R = self._build_Dtilde(dn1_R, dn2_R, dn0_R)
             for gamma in range(3):
-                _, _, U_A = self.ci._cpci_ints(Perturbation('vecpot', gamma))
+                _, _, U_A = self._cpci_ints(Perturbation('vecpot', gamma))
                 Icphi[la, gamma] = c('pq,pq->', D_tilde_R, U_A)
         return Icphi
 
@@ -630,7 +596,7 @@ class CIderiv(CorrelatedDerivs):
         for la in range(3 * natom):
             A, beta_ = divmod(la, 3)
             pR = Perturbation('nuclear', (A, beta_))
-            _, _, U_R = self.ci._cpci_ints(pR)
+            _, _, U_R = self._cpci_ints(pR)
             half_S = np.asarray(self.ci.derivatives.overlap_half(A)[beta_])
             Ur_eff = U_R + half_S.T
             for gamma in range(3):
@@ -647,18 +613,18 @@ class CIderiv(CorrelatedDerivs):
         I_pp = np.zeros((3 * natom, 3), dtype=complex)
         for la in range(3 * natom):
             A, beta_ = divmod(la, 3)
-            _, _, U_R = self.ci._cpci_ints(Perturbation('nuclear', (A, beta_)))
+            _, _, U_R = self._cpci_ints(Perturbation('nuclear', (A, beta_)))
             half_S = np.asarray(self.ci.derivatives.overlap_half(A)[beta_])
             Ur_eff = U_R + half_S.T
             for gamma in range(3):
-                _, _, U_A = self.ci._cpci_ints(Perturbation('vecpot', gamma))
+                _, _, U_A = self._cpci_ints(Perturbation('vecpot', gamma))
                 I_pp[la, gamma] = c('pq,pq->', D_pq, U_A.T @ Ur_eff)
         return I_pp
 
     def velocity_dipole_derivatives(self, gauge: str = 'canonical') -> np.ndarray:
         """CISD correlation velocity-gauge APT, shape (natom, 3, 3): -2 times the four overlap
         blocks with the correlation 1-PDM in Iphiphi. The SCF reference and the Z_A delta nuclear
-        term are supplied by the pycc.apt(gauge='velocity') facade, matching MPderiv's contract."""
+        term are supplied by the pycc.apt(gauge='velocity') facade, matching MPderiv."""
         natom = self.ci.ref.molecule().natom()
         overlap_total = (self.compute_Icc_VG_APT() + self.compute_Icphi_VG_APT() + self.compute_Iphic_VG_APT() + self.compute_Iphiphi_VG_APT())
         return (-2.0 * overlap_total).real.reshape(natom, 3, 3)
