+++ pycc/tests/test_080_cisd_vg_apt.py
@@ -1,5 +1,5 @@
 """
-CISD velocity-gauge atomic polar tensors - CIwfn.velocity_dipole_derivatives()\n(the correlation block) through the pycc.apt(gauge='velocity') facade: the\nwave-function-overlap formulation, -2 (Icc + Icphi + Iphic + Iphiphi) with the\nvector-potential (linear momentum) CPCI response.
+CISD velocity-gauge atomic polar tensors - pycc.apt(pycc.CIderiv(ci), gauge='velocity') facade. velocity_dipole_derivatives() is CISD-specific (a wave-function-overlap construction the base does not provide), -2*(Icc+Icphi+Iphic+Iphiphi) with the vector-potential CPCI response.
 
 Reference generated by the standalone validated CISD-VCD
 implementation (cisd_vcd.py / cisd_apt_lg.py / cisd_analytic_hessian.py;
@@ -8,7 +8,8 @@ itself validated independently: the analytic Hessian against Psi4/Gaussian
 finite differences (~1e-6), the AATs against the determinant-based apyib
 reference (B. Shumberger; ~1e-11 a.u.), and the LG-APT (Z-vector route)
 against finite differences of the analytic dipole.  The CIwfn port
-reproduces these values to ~1e-10.  The references are TOTAL tensors, so
+reproduces these values to ~1e-10; the test tolerance (1e-6) absorbs the
+printed precision of the reference.  The references are TOTAL tensors, so
 the tests exercise the full facade decomposition (nuclear + HF reference +
 CISD correlation).  Spatial orbitals, all-electron only.
 """
@@ -35,8 +36,7 @@ VG_APT_REF = np.array([
 
 
 def test_cisd_vg_apt_vs_reference(cisd_h2o2):
-    """The facade total (via CIderiv.velocity_dipole_derivatives(), migrated
-    verbatim from the retired on-CIwfn overlap code) reproduces the
-    standalone validated reference, every element."""
-    P = np.asarray(pycc.apt(cisd_h2o2(), gauge='velocity').total).reshape(-1, 3)
+    """The facade total reproduces the standalone validated reference, every
+    element."""
+    P = np.asarray(pycc.apt(pycc.CIderiv(cisd_h2o2()), gauge='velocity').total).reshape(-1, 3)
     assert np.max(np.abs(P - VG_APT_REF)) < 1e-9, np.max(np.abs(P - VG_APT_REF))
